requirements.txt
Bokeh==2.2.1
django==2.2.19
django_crontab==0.7.1
django-cors-headers==3.2.1
django-filter
docutils==0.16
future
jsonpickle
lxml==4.5.0
markdown
matplotlib
numpy
OWSLib>=0.22
pandas
#psycopg2
psycopg2-binary
pyzip
redis
requests
#watts_rsp
